chore: restructure the order summary component
