+++ packages/app-elements/src/ui/resources/ResourceOrderSummary/index.ts
@@ -1,0 +1,4 @@
+export {
+  ResourceOrderSummary,
+  type ResourceOrderSummaryProps
+} from './ResourceOrderSummary'
